featureguard.go
package featureguard

import (
	"context"

	"github.com/redis/go-redis/v9"
)

var (
	defaultRedisKeyPrefix = "featureguard:"
)

// FeatureGuard represents a feature flags management system using Redis as the storage backend.
type FeatureGuard struct {
	client    *redis.Client
	keyPrefix string
}

// NewFeatureGuard initializes a new FeatureGuard instance with the provided Redis client connection and key prefix.
func NewFeatureGuard(redisClient *redis.Client, redisKeyPrefix string) *FeatureGuard {
	if redisKeyPrefix == "" {
		redisKeyPrefix = defaultRedisKeyPrefix
	}

	return &FeatureGuard{
		client:    redisClient,
		keyPrefix: redisKeyPrefix,
	}
}

// EnableFeature enables the specified feature flag.
func (fg *FeatureGuard) EnableFeature(ctx context.Context, feature string) error {
	return nil
}

// DisableFeature disables the specified feature flag.
func (fg *FeatureGuard) DisableFeature(ctx context.Context, feature string) error {
	return nil
}

// ToggleFeature toggles the state of the specified feature flag.
func (fg *FeatureGuard) ToggleFeature(ctx context.Context, feature string) error {
	return nil
}

// IsFeatureEnabled checks if the specified feature flag is enabled
func (fg *FeatureGuard) IsFeatureEnabled(ctx context.Context, feature string) (bool, error) {
	return false, nil
}
